requirements.txt
flask
gunicorn
flask_sqlalchemy
flask_login
python-dotenv
psycopg2
bcrypt
flask_socketio